requirements.txt
# Core dependencies for Milestone 1
requests>=2.31.0
pydantic>=2.5.0
tqdm>=4.66.0

# LangChain ecosystem (for future milestones, but required now)
langchain>=0.1.0
langgraph>=0.0.26
langsmith>=0.0.87

# Hugging Face Hub (for future milestones)
huggingface-hub>=0.20.0

# OpenAI API (for future milestones)
openai>=1.12.0

# Text processing
tiktoken>=0.5.0
pymupdf4llm>=0.0.8

# Database (optional for now)
psycopg>=3.1.0

# UI framework (for future milestones)
gradio>=4.0.0

# Graph library
grandalf>=0.6

# Environment management
python-dotenv>=1.0.0

